This represents more than 3 months worth of changes that make KolourPaint a lot better.
However, KolourPaint still needs a lot of work to just work.

The merge took longer than expected as I had to fight SVN quite a bit :(

I'm worried about libkolourpaint_lgpl and symbol visibility as I saw some
KDE_EXPORT stuff.  Could someone who knows more about this please look into it?
Thanks in advance!


Forward ports from branches/KDE/3.5/:

* Scanning fixes

* Save local files atomically

* Implement rudimentary global session management

* Drop overly-spammed and unchangeable kolourpaint-support@lists.sourceforge.net

* "File / Open Recent" fixes

* CTRL+C'ing a text box also places the text in the middle-mouse-button
  clipboard, in lieu of being able to highlight the text to do this

* Change minimum allowed zoom level for the grid from 600% to 400%


Unique changes (not in branches/KDE/3.5/):

* Remove individual "Thanks To".  Unfortunately, it became unrealistic to keep track of everyone so I give up.  It made me feel really bad to have an incomplete list.

* Change rotate menu items to be consistent with Digikam; "Image / Rotate..." goes back to CTRL+R 
* Add, to the print dialog, a choice between printing the image at the
  top-left of the page or at the center (this was previously a hidden
  configuration option, which is now ignored) [Bug #133481]

* Add Zoom Tool

* Add "Image / Drag Opaque"

* kpTool::mouseReleaseEvent() finally calls draw() before endDraw().

  This means that a quick drag of the mouse immediately before releasing the
  mouse button is no longer ignored in tools that implement just draw()
  and not endDraw() [not sure there are such tools anymore though].
    
* Add Tone Enhance effect (Mike Gashler)

* Add "File / Properties..."

* Decouple from kpMainWindow using kp*Environment facades

* Massive selection code refactor
  - Support creating text selections that are only borders (contain no text
    and will not mutate document-is-modified state) i.e. you will be able to
    drag out text boxes and cancel them without affecting the document-is-modified flag.
    [this currently does not work due to lack of support from kpToolText]

* Identify a number of KolourPaint/KDE3 bugs (added TODOs)

* Make all selection tools work

* Rename kpTempPixmap -> kpTempImage

* Split kpCommandSize out of kpPixmapFX

* Much other refactoring

* Configurable color palette and "Colors" menu

* Animate the Color Similarity Tool Bar Item, to highlight the existence
  of the feature
  - And make the configuration more accessible
    - Also add "Image / Draw With Color Similarity" to duplicate the tool
      bar item

* Add support library containing LGPL code derived from elsewhere




svn path=/trunk/KDE/kdegraphics/kolourpaint/; revision=692114

+++ dialogs/kpDocumentSaveOptionsPreviewDialog.h
@@ -62,7 +62,7 @@ signals:
     void finished ();
 
 public slots:
-    void setFilePixmapAndSize (const QPixmap &filePixmap, int fileSize);
+    void setFilePixmapAndSize (const QPixmap &filePixmap, qint64 fileSize);
     void updatePixmapPreview ();
 
 protected:
@@ -72,7 +72,7 @@ protected:
 
 protected:
     QPixmap *m_filePixmap;
-    int m_fileSize;
+    qint64 m_fileSize;
 
     kpResizeSignallingLabel *m_filePixmapLabel;
     QLabel *m_fileSizeLabel;
